chore: rename zustand folder to store
